Fixed 32 second error

+++ heartbeat/heartbeat.ino
@@ -8,7 +8,7 @@ void loop() {
   digitalWrite(13, HIGH);
   delay(50);
   digitalWrite(13, LOW);
-  int x = millis();
+  long x = millis();
   Serial.print(x/1000);
   Serial.print(" sec have elapsed. millis = ");
   Serial.println(x);
